Make string.format checks a little stricter
Technically a breaking change, but hopefully not in practice.

+++ src/main/java/cc/tweaked/internal/string/NumberParser.java
@@ -187,6 +187,6 @@ public final class NumberParser {
 			exponent += givenExponent;
 		}
 
-		return Math.scalb(result, exponent);
+		return Math.scalb((double) result, exponent);
 	}
 }
